chore: move api tokens context to its edit view
